move tests for metrics to dedicated unit test class
