+++ service/src/main/java/bio/terra/profile/service/profile/flight/create/CreateProfileVerifyAccountStep.java
@@ -14,7 +14,7 @@ import com.google.iam.v1.TestIamPermissionsResponse;
 import java.util.List;
 
 /** Step to verify the user has access to a GCP profile's billing account. */
-record CreateProfileVerifyAccountStep(
+public record CreateProfileVerifyAccountStep(
     CrlService crlService, BillingProfile profile, AuthenticatedUserRequest user) implements Step {
 
   public static List<String> PERMISSIONS_TO_TEST = List.of("billing.resourceAssociations.create");
